Rename refactoring, no change in functionality.
