+++ LibraryManagementSystem/Repositories/InMemoryRepository.cs
@@ -1,7 +1,12 @@
-﻿using LibraryManagementSystem.Interfaces;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using LibraryManagementSystem.Interfaces;
 using LibraryManagementSystem.Models;
 
-namespace LibraryManagementSystem.Respositories
+namespace LibraryManagementSystem.Repositories
 {
     public class InMemoryRepository<T> : IRepository<T> where T : IEntity
     {
@@ -37,5 +42,4 @@ namespace LibraryManagementSystem.Respositories
             return Exists(id) ? _entities[id] : default(T);
         }
     }
-    
 }
